nput validation added for numeric input; support for decimal numbers (dot) implemented

+++ src/Calculator.Calculations/Calculations.cs
@@ -1,4 +1,5 @@
 ﻿using Serilog;
+using System;
 
 namespace Calculator.Calculations
 {
@@ -6,28 +7,28 @@ namespace Calculator.Calculations
     {
 
         
-        public static int Sum(int a, int b)
+        public static double Sum(double a, double b)
         {
-            int result = a + b;
+            double result = a + b;
             Log.Information("Součet {A} + {B} = {Result}", a, b, result);
             return result;
         }
 
-        public static int Difference(int a, int b)
+        public static double Difference(double a, double b)
         {
-            int result = a - b;
+            double result = a - b;
             Log.Information("Rozdíl {A} - {B} = {Result}", a, b, result);
             return result;
         }
 
-        public static int Multiply(int a, int b)
+        public static double Multiply(double a, double b)
         {
-            int result = a * b;
+            double result = a * b;
             Log.Information("Součin {A} * {B} = {Result}", a, b, result);
             return result;
         }
 
-        public static double Division(int a, int b)
+        public static double Division(double a, double b)
         {
             if (b == 0)
             {
